requirements_20240220.txt
aggdraw==1.3.16
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.4
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
clr-loader==0.2.6
colorama==0.4.6
Deprecated==1.2.14
docopt==0.6.2
eng-to-ipa==0.0.2
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.31.1
jaconv==0.3.4
Janome==0.4.2
lazy_loader==0.3
lxml==4.9.2
multidict==6.0.4
networkx==3.1
numpy==1.24.2
openai==0.28.1
packaging==23.1
pefile==2023.2.7
Pillow==9.4.0
psd-tools==1.9.28
psutil==5.9.4
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.6
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pykakasi==2.2.1
python-dotenv==1.0.0
pythonnet==3.0.3
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.2
PyYAML==6.0
requests==2.28.2
scikit-image==0.21.0
scipy==1.11.1
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
tifffile==2023.7.18
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.18.1
websocket-client==1.7.0
websockets==10.4
wrapt==1.16.0
yarl==1.8.2
